Update game server service protobuf

+++ app/game/server/model/server.go
@@ -1,9 +1,9 @@
 package model
 
 type Info struct {
-	ServerID int64  `gorm:"primary_key;unique;not null" json:"server_id"`
-	ModID    int64  `gorm:"not null" json:"mod_id"`
-	GameID   int64  `gorm:"not null" json:"game_id"`
+	ServerID int32  `gorm:"primary_key;unique;not null" json:"server_id"`
+	ModID    int32  `gorm:"not null" json:"mod_id"`
+	GameID   int32  `gorm:"not null" json:"game_id"`
 	Rcon     string `gorm:"not null" json:"rcon"`
 	Hostname string `gorm:"not null" json:"hostname"`
 	Address  string `gorm:"not null;unique" json:"address"`
